pyproject.toml
[project]
name = "demo-ca"
version = "0.7.0"
description = "Clean Architecture domain business layer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["django>=5.2.12", "djangorestframework>=3.16.1"]

[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["domain"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4"]
ignore = ["E501"]

[tool.pyright]
typeCheckingMode = "standard"
include = ["domain"]

[dependency-groups]
dev = ["pyright>=1.1.408", "pytest>=9.0.2", "ruff>=0.15.6"]
